# swagger.yaml
openapi: 3.0.3
info:
  title: Super API gerenciadora de filmes
  description: API para gerenciamento de filmes, contendo controle completo de elenco, equipe técnica, pessoas, premiações e mídias associadas.
  contact:
    email: gabriel4321uo@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080/v1/senai/locadora
paths:
...
  /pessoa/:
    get:
      tags:
        - pessoa
      summary: Rota usada para obter todos as pessoas.
      description: Rota usada para obter todos as pessoas.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      pessoa:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pessoa'
        '404':
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
    post:
      tags:
        - pessoa
      summary: Rota usada para inserir uma nova pessoa.
      description: Rota usada para inserir uma nova pessoa.
      responses:
        '201':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 201
                  response:
                    type: object
                    properties:
                      pessoa:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pessoa'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pessoa'
  /pessoa/:id:
    get:
      tags:
        - pessoa
      summary: Rota usada para obter uma pessoa pelo ID.
      description: Rota usada para obter uma pessoa pelo ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id da pessoa a pesquisar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      pessoa:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pessoa'
    put:
      tags:
        - pessoa
      summary: Rota usada para atualizar uma pessoa.
      description: Rota usada para atualizar uma pessoa.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id da pessoa a atualizar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      pessoa:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pessoa'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pessoa'
    delete:
      tags:
        - pessoa
      summary: Rota usada para deletar uma pessoa.
      description: Rota usada para deletar uma pessoa.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id da pessoa a deletar
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SUCCESS_DELETED_ITEM'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
...
  /personagem/:
    get:
      tags:
        - personagem
      summary: Rota para obter todos os personagens.
      description: Rota para obter todos os personagens.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      personagem:
                        type: array
                        items:
                          $ref: '#/components/schemas/Personagem'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    post:
      tags:
        - personagem
      summary: Rota para criar um personagem
      description: Rota para criar um personagem
      responses:
        '201':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 201
                  response:
                    type: object
                    properties:
                      personagem:
                        type: array
                        items:
                          $ref: '#/components/schemas/Personagem'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Personagem'
  /personagem/:id:
    get:
      tags:
        - personagem
      summary: Rota para obter um personagem pelo id.
      description: Rota para obter um personagem pelo id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do personagem a obter.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      personagem:
                        type: array
                        items:
                          $ref: '#/components/schemas/Personagem'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    put:
      tags:
        - personagem
      summary: Rota para atualizar um personagem
      description: Rota para atualizar um personagem
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do personagem a atualizar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      personagem:
                        type: array
                        items:
                          $ref: '#/components/schemas/Personagem'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Personagem'
    delete:
      tags:
        - personagem
      summary: Rota do personagem a deletar
      description: Rota do personagem a deletar
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SUCCESS_DELETED_ITEM'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
...
components:
  schemas:
    ERROR_BAD_REQUEST:
      type: object
      properties:
        status:
          type: boolean
          description: status false
        status_code:
          type: integer
          format: int64
          description: '400'
        message:
          type: string
          description: Não foi possivel processar a requisição devido a erros de entrada de dados 😣.
        image:
          type: string
          description: https://http.dog/400.jpg
    SUCCESS_DELETED_ITEM:
      type: object
      properties:
        status:
          type: boolean
          description: 'true'
        status_code:
          type: integer
          format: int64
          description: '200'
        message:
          type: string
          description: Item excluido com sucesso 🥳
        image:
          type: string
          description: https://http.dog/200.jpg
    ERROR_INTERNAL_SERVER_MODEL:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '500'
        message:
          type: string
          description: Não foi possível processar a requisição devido a um erro interno no servidor. 🤖 [MODEL].
        image:
          type: string
          description: https://http.dog/500.jpg
    ERROR_INTERNAL_SERVER_CONTROLLER:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '500'
        message:
          type: string
          description: Não foi possível processar a requisição devido a um erro interno no servidor. 🤖 [CONTROLLER].
        image:
          type: string
          description: https://http.dog/500.jpg
    ERROR_CONTENT_TYPE:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '415'
        message:
          type: string
          description: Não foi possível processar a requisição, pois o formato de dados ecaminhado não é suportado pelo servidor, apenas deve ser utilizado JSON. 🤖
        image:
          type: string
          description: https://http.dog/415.jpg
    ERROR_NOT_FOUND:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '404'
        message:
          type: string
          description: Não foram encontrados dados para retorno. 🗺️❌
        image:
          type: string
          description: https://http.dog/404.jpg
...
    Pessoa:
      required:
        - nome_nascimento
        - data_nascimento
      type: object
      properties:
        nome_nascimento:
          type: string
          description: Nome da pessoa a inserir.
        data_nascimento:
          type: string
          format: date
          description: Data de nascimento da pessoa a inserir.
        data_falecimento:
          type: string
          format: date
          description: Data de falecimento da pessoa a inserir.
        ano_inicio_carreira:
          type: integer
          format: int64
          description: Ano de inicio de carreira da pessoa a inserir.
        biografia:
          type: string
          description: Biografia da pessoa a inserir.
        altura:
          type: number
          description: Altura da pessoa a inserir.
        cidade_natal:
          type: string
          description: Nome da cidade natal da pessoa a inserir.
        id_sexo:
          type: integer
          format: int64
          description: Id do sexo da pessoa a inserir.
        curiosidades:
          type: array
          description: Array que guarda ids com curiosidades sobre a pessoa a inserir.
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
                description: id da curiosidade.
            required:
              - id
        premios:
          type: array
          description: Premios recebidos pela pessoa a inserir.
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
                description: Id do premio recebido.
            required:
              - id
        paises_origem:
          type: array
          description: Paises de origem da pessoa a inserir.
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
                description: Id do pais.
            required:
              - id
        ocupacoes:
          type: array
          description: Ocupações profissionais da pessoa a inserir.
          items:
            type: object
            properties:
              id:
                type: integer
                format: int64
                description: id da ocupação.
            required:
              - id
        fotos:
          type: array
          description: Fotos da pessoa a inserir.
          items:
            type: object
            properties:
              foto:
                type: string
                description: url da foto a inserir.
            required:
              - foto
        nomes_artisticos:
          type: array
          description: Nomes artisticos da pessoa a inserir.
          items:
            type: object
            properties:
              nome:
                type: string
                description: Nome artisco da pessoa
            required:
              - nome
...
    Personagem:
      required:
        - nome
        - descricao
      type: object
      properties:
        nome:
          type: string
          description: Nome do personagem a inserir.
        descricao:
          type: string
          description: Descrição do personagem a inserir.
        fotos:
          type: array
          description: Fotos do personagem a inserir.
          items:
            type: object
            properties:
              foto:
                type: string
                description: url da foto do personagem.
            required:
              - foto
...